backend/requirements.txt
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.0.0
python-jose>=3.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pytest>=9.0.0
